requirements.txt
Flask==3.0.3
selenium-wire==5.1.0
webdriver-manager==4.0.2
requests==2.32.3
beautifulsoup4==4.12.3
blinker==1.7.0
yt-dlp>=2023.12.30
Flask-WTF>=1.2.1
python-dotenv>=1.0.1
werkzeug>=3.0.0